test_bot_full/requirements.txt
aiogram==3.18.0
python-dotenv==1.0.1
requests==2.32.3
apscheduler==3.11.0
openai==1.66.3
gspread==6.0.2
asyncpg>=0.27.0
